backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
authlib
python-jose[cryptography]
passlib[bcrypt]
apscheduler
httpx
python-multipart
